requirements.txt
tinygrad>=0.10.0
numpy
pillow
